Revert "WEBRTC-2550: Fix background activity launch blocked issue and make notification click accept call"
This reverts commit bf479af723fca37b0ec92ebd57485d00df5f09f4.
